requirements.txt
Flask==2.2.2
Flask_WTF==1.0.1
WTForms==3.0.1

Flask-QRcode
Flask
Jinja2
MarkupSafe
Pillow
Werkzeug
click
importlib-metadata
itsdangerous
pylightning
qrcode
zipp
wheel